Display Module Screen Blackout Detection
The following scripts used Python and OpenCV to detect screen blackout
in display modules, where the input is a video file (ex. mp4) of the
screen, and output is a text file reporting frames where blackouts
occurred. This is computed by comparing pixel values within the region
of interest against a determined threshold.